Fix LowList Poster image selection logic.
Special listings that contain more than one show or season (such as
RecentlyAddedEpisodes and All Seasons) don't set Container.SeasonThumb
for obvious reasons.  They should instead set ListItem.Property(SeasonThumb)
but XBMC appears to currently only set that for DialogVideoInfo.

The new logic attempts to use ListItem.Property(SeasonThumb) if it's in
either of those views, and fall back to Container.TvshowThumb (if in
All Seasons) or to tvshow.jpg (if in RecentlyAddedEpisodes).

This should make it magically start working as intended once XBMC fills
in the SeasonThumb listitem property, but for now use reasonable defaults.