Remove extrafanart from musicvideos, addons, and sets.
They're not included in the 'spec' on the Kodi wiki anyway.

This only partially fixes Issue #303.  If the user enables extrafanart and
has an item in her library that does not have extra fanart, it will display
the extrafanart from the previously-viewed item, which is wrong.

But I don't know what to do about this.